src/test/easy/containsDuplicate/ContainsDuplicateTest.java
package easy.containsDuplicate;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.*;

class ContainsDuplicateTest {

    public static Stream<Arguments> data() {
        return Stream.of(
                Arguments.of(new int[]{1, 2, 3, 1}, true),
                Arguments.of(new int[]{1, 2, 3, 4}, false),
                Arguments.of(new int[]{1, 1, 1, 3, 3, 4, 3, 2, 4, 2}, true)
        );
    }

    @ParameterizedTest
    @MethodSource("data")
    void test(int[] arr, boolean answer) {
        //given
        ContainsDuplicate test = new ContainsDuplicate();
        //when
        boolean result = test.containsDuplicate(arr);
        //then
        assertEquals(result, answer);
    }
}